moved helpers, fixed issue from row major -> cartesian again
