+++ src/brain_input/vision/peripheral_segmentation.rs
@@ -1,4 +1,4 @@
-use super::single_frame::ImageFrame;
+use super::ImageFrame;
 use crate::error::DataProcessingError;
 use super::descriptors::*;
 use crate::cortical_data::CorticalID;
@@ -367,175 +367,8 @@ impl SegmentedVisionFrame {
 }
 
 
-// region helpers
+// region internal helpers
 
-/// Properties defining the center region of a segmented vision frame
-///
-/// This structure defines the coordinates and size of the central region
-/// in a normalized coordinate space (0.0 to 1.0).
-#[derive(PartialEq, Clone, Copy)]
-pub struct SegmentedVisionCenterProperties {
-    /// Center point coordinates in normalized space (0.0-1.0), from the top left
-    center_coordinates_normalized_yx: (f32, f32), // Scaled from 0 to 1
-    /// Size of the center region in normalized space (0.0-1.0)
-    center_size_normalized_yx: (f32, f32), // ditto
-}
-
-impl SegmentedVisionCenterProperties {
-    /// Creates a new instance with normalized center coordinates and size
-    ///
-    /// # Arguments
-    ///
-    /// * `center_coordinates_normalized` - Center point as (x, y) in normalized space (0.0-1.0)
-    /// * `center_size_normalized` - Size as (width, height) in normalized space (0.0-1.0)
-    ///
-    /// # Returns
-    ///
-    /// * `Result<SegmentedVisionCenterProperties, &'static str>` - Created instance or error message
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if any coordinate or size value is outside the range [0.0, 1.0]
-    pub fn new_row_major_where_origin_top_left(center_coordinates_normalized_yx_rm_tl: (f32, f32), center_size_yx_normalized: (f32, f32)) -> Result<SegmentedVisionCenterProperties, DataProcessingError> {
-        if center_coordinates_normalized_yx_rm_tl.0 < 0.0 || center_coordinates_normalized_yx_rm_tl.1 < 0.0 || center_coordinates_normalized_yx_rm_tl.0 > 1.0 || center_coordinates_normalized_yx_rm_tl.1 > 1.0 {
-            return Err(DataProcessingError::InvalidInputBounds("Coordinates are to be normalized and must be between 0 and 1!".into()))
-        }
-        if center_size_yx_normalized.0 < 0.0 || center_size_yx_normalized.1 < 0.0 || center_size_yx_normalized.0 > 1.0 || center_size_yx_normalized.1 > 1.0 {
-            return Err(DataProcessingError::InvalidInputBounds("Central vision size is to be normalized and must be between 0 and 1!".into()))
-        }
-        Ok(SegmentedVisionCenterProperties {
-            center_coordinates_normalized_yx: center_coordinates_normalized_yx_rm_tl,
-            center_size_normalized_yx: center_size_yx_normalized,
-        })
-    }
-
-    pub fn cartesian_where_origin_bottom_left(center_coordinates_normalized_cartesian_yx: (f32, f32), center_size_normalized_yx: (f32, f32)) -> Result<SegmentedVisionCenterProperties, DataProcessingError> {
-        SegmentedVisionCenterProperties::new_row_major_where_origin_top_left(
-            (1.0 - center_coordinates_normalized_cartesian_yx.0, center_coordinates_normalized_cartesian_yx.1),
-            center_size_normalized_yx)
-    }
-
-    pub fn create_default_centered() -> SegmentedVisionCenterProperties {
-        SegmentedVisionCenterProperties::new_row_major_where_origin_top_left((0.5, 0.5), (0.5, 0.5)).unwrap()
-    }
-
-    /// Returns an instance of CornerPoints that defines the region of the center region by pixel index
-    ///
-    /// # Arguments
-    ///
-    /// * `source_frame_width_height` - The total resolution (width, height) of the source frame in pixels
-    ///
-    /// # Returns
-    ///
-    /// * `Result<CornerPoints, &'static str>` - Corner points that define the center region or an error message
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the source width or height is less than 3 pixels
-    pub fn calculate_pixel_coordinates_of_center_corners(&self, source_frame_width_height: (usize, usize)) -> Result<CornerPoints, DataProcessingError> {
-        if source_frame_width_height.0 < 3 || source_frame_width_height.1 < 3 {
-            return Err(DataProcessingError::InvalidInputBounds("Source resolution must be 3 pixels or greater in the X and Y directions!".into()));
-        }
-        let source_frame_width_height_f: (f32, f32) = (source_frame_width_height.0 as f32, source_frame_width_height.1 as f32);
-        let center_size_normalized_half_yx: (f32, f32) = (self.center_size_normalized_yx.0 / 2.0, self.center_size_normalized_yx.1 / 2.0);
-
-        // We use max / min to ensure that there is always a 1 pixel buffer along all edges for use in peripheral vision (since we cannot use a resolution of 0)
-        let bottom_pixel: usize = cmp::min(source_frame_width_height.0 - 1,
-                                           ((self.center_coordinates_normalized_yx.0 + center_size_normalized_half_yx.0) * source_frame_width_height_f.1).ceil() as usize);
-        let top_pixel: usize = cmp::max(1,
-                                        (( self.center_coordinates_normalized_yx.0 - center_size_normalized_half_yx.0) * source_frame_width_height_f.1).floor() as usize);
-        let left_pixel: usize = cmp::max(1,
-                                         ((self.center_coordinates_normalized_yx.1 - center_size_normalized_half_yx.1) * source_frame_width_height_f.0).floor() as usize);
-        let right_pixel: usize = cmp::min(source_frame_width_height.0 - 1,
-                                          (( self.center_coordinates_normalized_yx.1 + center_size_normalized_half_yx.1) * source_frame_width_height_f.0).ceil() as usize);
-
-        let corner_points: CornerPoints = CornerPoints::new_from_row_major(
-            (bottom_pixel, left_pixel),
-            (top_pixel, right_pixel)
-        )?;
-        Ok(corner_points)
-    }
-
-
-}
-
-/// Target resolutions for each of the nine segments in a segmented vision frame
-///
-/// This structure stores the desired output resolution for each of the segments
-/// in a grid arrangement (3x3): corners, edges, and center.
-#[derive(PartialEq, Clone, Copy)]
-pub struct SegmentedVisionTargetResolutions {
-    /// Resolution for lower-left segment as (width, height)
-    lower_left: (usize, usize),
-    /// Resolution for middle-left segment as (width, height)
-    middle_left: (usize, usize),
-    /// Resolution for upper-left segment as (width, height)
-    upper_left: (usize, usize),
-    /// Resolution for upper-middle segment as (width, height)
-    upper_middle: (usize, usize),
-    /// Resolution for upper-right segment as (width, height)
-    upper_right: (usize, usize),
-    /// Resolution for middle-right segment as (width, height)
-    middle_right: (usize, usize),
-    /// Resolution for lower-right segment as (width, height)
-    lower_right: (usize, usize),
-    /// Resolution for lower-middle segment as (width, height)
-    lower_middle: (usize, usize),
-    /// Resolution for center segment as (width, height)
-    center: (usize, usize),
-}
-
-impl SegmentedVisionTargetResolutions {
-    /// Creates a new SegmentedVisionTargetResolutions with specified resolutions for each segment
-    ///
-    /// # Arguments
-    ///
-    /// * Nine pairs of (width, height) values, one for each segment
-    ///
-    /// # Returns
-    ///
-    /// * `Result<SegmentedVisionTargetResolutions, &'static str>` - Created instance or error message
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if any dimension is zero
-    pub fn new(
-        lower_left: (usize, usize),
-        middle_left: (usize, usize),
-        upper_left: (usize, usize),
-        upper_middle: (usize, usize),
-        upper_right: (usize, usize),
-        middle_right: (usize, usize),
-        lower_right: (usize, usize),
-        lower_middle: (usize, usize),
-        center: (usize, usize),
-    ) -> Result<SegmentedVisionTargetResolutions, DataProcessingError> {
-        if lower_left.0 == 0 || lower_left.1 == 0 || middle_left.0 == 0 || middle_left.1 == 0 || upper_left.0 == 0 || upper_left.1 == 0 || upper_middle.0 == 0 || upper_middle.1 == 0 || // Yandre-dev moment
-            upper_right.0 == 0 || upper_right.1 == 0 || middle_right.0 == 0 || middle_right.1 == 0 || lower_right.0 == 0 || lower_right.1 == 0 || lower_middle.0 == 0 || lower_middle.1 == 0 ||
-            center.0 == 0 || center.1 == 0 {
-            return Err(DataProcessingError::InvalidInputBounds("Dimensions must exceed 0 for all segments on all axis!".into()));
-        }
-        Ok(SegmentedVisionTargetResolutions {
-            lower_left,
-            middle_left,
-            upper_left,
-            upper_middle,
-            upper_right,
-            middle_right,
-            lower_right,
-            lower_middle,
-            center,
-        })
-    }
-
-    pub fn create_with_same_sized_peripheral(center_width_height: (usize, usize), peripheral_width_height: (usize, usize)) -> Result<SegmentedVisionTargetResolutions, DataProcessingError> {
-        return SegmentedVisionTargetResolutions::new(peripheral_width_height, peripheral_width_height,
-                                                     peripheral_width_height, peripheral_width_height,
-                                                     peripheral_width_height, peripheral_width_height,
-                                                     peripheral_width_height, peripheral_width_height, center_width_height);
-    }
-
-}
 
 /// Stores the corner points for each segment of a segmented vision frame
 ///
